+++ src/Exception/Exception.php
@@ -1,0 +1,23 @@
+<?php
+/* (c) Anton Medvedev <anton@medv.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Deployer\Exception;
+
+class Exception extends \Exception
+{
+    /**
+     * Make message mandatory.
+     *
+     * @param string $message
+     * @param integer $code defaults to 0
+     * @param \Exception|null $previous
+     */
+    public function __construct($message, $code = 0, \Exception $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+    }
+}