+++ src/Exception/NonFatalException.php
@@ -7,17 +7,6 @@
 
 namespace Deployer\Exception;
 
-class NonFatalException extends \RuntimeException
+class NonFatalException extends Exception
 {
-    /**
-     * Make message mandatory.
-     *
-     * @param string $message
-     * @param integer $code defaults to 0
-     * @param \Exception|null $previous
-     */
-    public function __construct($message, $code = 0, \Exception $previous = null)
-    {
-        parent::__construct($message, $code, $previous);
-    }
 }