+++ recipe/magento2.php
@@ -4,27 +4,39 @@ namespace Deployer;
 require_once __DIR__ . '/common.php';
 
 // Configuration
-set('shared_files', ['app/etc/env.php', 'var/.maintenance.ip']);
-set('shared_dirs', ['var/log', 'var/backups', 'pub/media']);
-set('writable_dirs', ['var', 'pub/static', 'pub/media']);
-set('clear_paths', ['var/generation/*', 'var/cache/*']);
+set('shared_files', [
+    'app/etc/env.php',
+    'var/.maintenance.ip',
+]);
+set('shared_dirs', [
+    'var/log',
+    'var/backups',
+    'pub/media',
+]);
+set('writable_dirs', [
+    'var',
+    'pub/static',
+    'pub/media',
+]);
+set('clear_paths', [
+    'var/generation/*',
+    'var/cache/*',
+]);
 
 // Tasks
-desc('Enable all modules'); task('magento:enable', function () {
+desc('Enable all modules');
+task('magento:enable', function () {
     run("{{bin/php}} {{release_path}}/bin/magento module:enable --all");
-    writeln("Modules enabled");
 });
 
 desc('Compile magento di');
 task('magento:compile', function () {
     run("{{bin/php}} {{release_path}}/bin/magento setup:di:compile");
-    writeln("DI compiled");
 });
 
 desc('Deploy assets');
 task('magento:deploy:assets', function () {
     run("{{bin/php}} {{release_path}}/bin/magento setup:static-content:deploy");
-    writeln("Static assets deployed");
 });
 
 desc('Enable maintenance mode');
@@ -41,30 +53,12 @@ desc('Upgrade magento database');
 task('magento:upgrade:db', function () {
     run("{{bin/php}} {{release_path}}/bin/magento setup:db-schema:upgrade");
     run("{{bin/php}} {{release_path}}/bin/magento setup:db-data:upgrade");
-    writeln("Database upgraded");
 });
 
 desc('Flush Magento Cache');
 task('magento:cache:flush', function () {
     run("{{bin/php}} {{release_path}}/bin/magento cache:flush");
-    writeln("Magento cache flushed");
 });
-
-desc('Deploy your project');
-task('deploy', [
-    'deploy:prepare',
-    'deploy:lock',
-    'deploy:release',
-    'deploy:update_code',
-    'deploy:shared',
-    'deploy:writable',
-    'deploy:vendors',
-    'deploy:clear_paths',
-    'deploy:symlink',
-    'deploy:unlock',
-    'cleanup',
-    'success'
-]);
 
 desc('Magento2 deployment operations');
 task('deploy:magento', [
@@ -77,4 +71,19 @@ task('deploy:magento', [
     'magento:maintenance:disable'
 ]);
 
-after('deploy:clear_paths', 'deploy:magento');
+desc('Deploy your project');
+task('deploy', [
+    'deploy:prepare',
+    'deploy:lock',
+    'deploy:release',
+    'deploy:update_code',
+    'deploy:shared',
+    'deploy:writable',
+    'deploy:vendors',
+    'deploy:clear_paths',
+    'deploy:magento',
+    'deploy:symlink',
+    'deploy:unlock',
+    'cleanup',
+    'success'
+]);