+++ CHANGELOG.md
@@ -49,5 +49,6 @@
 - Fixed typo3 recipe
 - Fixed remove of shared dir on first deploy
 
+
 ## v4.0.0
 🙄