+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added option `-L` to `setfacl` [#956](https://github.com/deployphp/deployer/pull/956)
 - Run `deploy:unlock` when deploy is fail [#958](https://github.com/deployphp/deployer/pull/958)
 - Now throw exception on duplicates in `shared_dirs`
+- `add()` now merges configuration options recursively
 
 ### Fixed
 - Fixed native ssh scp option