+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `testLocally` function (analog `test` fn)
 - Added `ConfigurationException`
 - Show message on file download
+- Added support for multiplexing for NativeSsh [#918](https://github.com/deployphp/deployer/pull/918)
 
 ### Changed
 - Server config `setPty` renamed to `pty` [#953](https://github.com/deployphp/deployer/pull/953)