+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `ConfigurationException`
 - Show message on file download
 - Added support for multiplexing for NativeSsh [#918](https://github.com/deployphp/deployer/pull/918)
+- Added GracefulShutdownException
 
 ### Changed
 - Server config `setPty` renamed to `pty` [#953](https://github.com/deployphp/deployer/pull/953)